+++ src/api/control/room.rs
@@ -38,14 +38,11 @@ pub enum RoomElement {
         credentials: String,
         on_leave: Option<CallbackUrl>,
         on_join: Option<CallbackUrl>,
-        #[serde(default)]
-        #[serde(with = "humantime_serde")]
+        #[serde(default, with = "humantime_serde")]
         idle_timeout: Option<Duration>,
-        #[serde(default)]
-        #[serde(with = "humantime_serde")]
+        #[serde(default, with = "humantime_serde")]
         reconnect_timeout: Option<Duration>,
-        #[serde(default)]
-        #[serde(with = "humantime_serde")]
+        #[serde(default, with = "humantime_serde")]
         ping_interval: Option<Duration>,
     },
 }